--- openeo_fastapi/client/processes/__init__.py
@@ -1,2 +1,0 @@
-from openeo_fastapi.client.processes.models import *
-from openeo_fastapi.client.processes.register import ProcessRegister
